+++ CHANGELOG.md
@@ -7,10 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-- Added optional dependencies for `databases`
+- Added optional dependencies for `databases` and `templates`
 - Added environment prefixes for database configuration
 - Added `templates` and `TemplatesConfig` to config
 - Added the `templates` function
+- Added support for `attrs` in type validation.
 
 ## [1.0.0-alpha7] - 2023-12-7